todo_list/pyproject.toml
[tool.poetry]
name = "todo-list"
version = "0.1.0"
description = ""
authors = ["usik-sviatoslav <usik.sviatoslav.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
drf-spectacular = "^0.27.2"
drf-social-oauth2 = "^3.0.1"
django-redis = "^5.4.0"
redis = "^5.0.7"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
django-filter = "^24.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
flake8 = "^7.1.0"
flake8-pyproject = "^1.2.3"
flake8-variables-names = "^0.0.6"
flake8-functions = "^0.0.8"
flake8-class-attributes-order = "^0.1.3"
black = "^24.4.2"
isort = "^5.13.2"
pytest = "^8.2.2"
pytest-mock = "^3.14.0"
pytest-django = "^4.8.0"
pytest-cov = "^5.0.0"
django-silk = "^5.1.0"

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120
exclude = [".env", ".git", "poetry.lock", ".pytest_cache", "__pycache__", "venv"]
max-complexity = 8
ignore = []

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings.dev"
python_files = ["tests.py", "test_*.py", "*_test.py"]
addopts = ["-x", "--ff", "--nf", "-W error", "--color=yes"]
cache_dir = "/tmp/.pytest_cache"

[tool.coverage.run]
branch = true
data_file = "htmlcov/.coverage"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
